AP-14 - Cameron James Fixed the find process and current process methods on repositoryService and changed the version numbering of the system.  The branch and processModelVersion for the name use "MAIN" for the initial insert. The version number is taken from the UI on insertion and placed on the processModelVersion.  Also added the wsdl back to the web app, wouldn't work without it and removed unused libraries in the extras folder.

+++ trunk/Apromore-Core/apromore-service/src/main/java/org/apromore/service/helper/ChangePropagator.java
@@ -81,7 +81,7 @@ public class ChangePropagator {
 
     private void createNewProcessModelVersion(ProcessModelVersion pmv, String rootFragmentUri, List<String> composingFragmentIds)
             throws ExceptionDao {
-        int versionNumber = pmv.getVersionNumber() + 1;
+        Double versionNumber = pmv.getVersionNumber() + 1;
         String versionName = VersionNameUtil.getNextVersionName(pmv.getVersionName());
         ProcessModelVersion pv = pSrv.addProcessModelVersion(pmv.getProcessBranch(), rootFragmentUri, versionNumber, versionName, 0, 0);
         fSrv.addProcessFragmentMappings(pv.getId(), composingFragmentIds);
